+++ src/main/java/learn/thyme/thyteams/user/CreateUserParameters.java
@@ -4,18 +4,25 @@ import java.time.LocalDate;
 
 public class CreateUserParameters {
     private final UserName userName;
+    /** Password in clear text. */
+    private final String password;
     private final Gender gender;
     private final LocalDate birthday;
     private final Email email;
     private final PhoneNumber phoneNumber;
 
-    public CreateUserParameters(UserName userName, Gender gender, LocalDate birthday,
+    public CreateUserParameters(UserName userName, String password, Gender gender, LocalDate birthday,
                                 Email email, PhoneNumber phoneNumber) {
         this.userName = userName;
+        this.password = password;
         this.gender = gender;
         this.birthday = birthday;
         this.email = email;
         this.phoneNumber = phoneNumber;
+    }
+
+    public String getPassword() {
+        return password;
     }
 
     public UserName getUserName() {