+++ src/main/java/learn/thyme/thyteams/user/User.java
@@ -2,17 +2,22 @@ package learn.thyme.thyteams.user;
 
 import io.github.wimdeblauwe.jpearl.AbstractVersionedEntity;
 
-import javax.persistence.Entity;
-import javax.persistence.EnumType;
-import javax.persistence.Enumerated;
-import javax.persistence.Table;
+import javax.persistence.*;
 import javax.validation.constraints.NotNull;
 import java.time.LocalDate;
+import java.util.Set;
 
 @Entity
 @Table(name = "tt_user")
 public class User extends AbstractVersionedEntity<UserId> {
 
+    @ElementCollection(targetClass = UserRole.class)
+    @Enumerated(EnumType.STRING)
+    @CollectionTable(name = "user_roles")
+    @Column(name = "role")  // name of the column where the enum value will be stored in 'user_roles' table
+    private Set<UserRole> roles;
+    @NotNull
+    private String password;
     @NotNull
     private UserName userName;
     @NotNull
@@ -31,14 +36,52 @@ public class User extends AbstractVersionedEntity<UserId> {
     protected User() {
     }
 
-    public User(UserId id, @NotNull UserName userName, @NotNull Gender gender, @NotNull LocalDate birthday,
-                @NotNull Email email, @NotNull PhoneNumber phoneNumber) {
+    private User(UserId id,
+                @NotNull Set<UserRole> roles,
+                @NotNull UserName userName,
+                @NotNull String password,
+                @NotNull Gender gender,
+                @NotNull LocalDate birthday,
+                @NotNull Email email,
+                @NotNull PhoneNumber phoneNumber) {
         super(id);
+        this.roles = roles;
         this.userName = userName;
+        this.password = password;
         this.gender = gender;
         this.birthday = birthday;
         this.email = email;
         this.phoneNumber = phoneNumber;
+    }
+
+    public static User createUser(UserId id,
+                                  UserName userName,
+                                  String encodedPassword,
+                                  Gender gender,
+                                  LocalDate birthday,
+                                  Email email,
+                                  PhoneNumber phoneNumber) {
+        return new User(id, Set.of(UserRole.USER), userName, encodedPassword, gender,
+                birthday, email, phoneNumber);
+    }
+
+    public static User createAdministrator(UserId id,
+                                           UserName userName,
+                                           String encodedPassword,
+                                           Gender gender,
+                                           LocalDate birthday,
+                                           Email email,
+                                           PhoneNumber phoneNumber) {
+        return new User(id, Set.of(UserRole.USER, UserRole.ADMIN), userName,
+                encodedPassword, gender, birthday, email, phoneNumber);
+    }
+
+    public Set<UserRole> getRoles() {
+        return roles;
+    }
+
+    public String getPassword() {
+        return password;
     }
 
     public UserName getUserName() {