+++ src/main/java/learn/thyme/thyteams/user/UserRole.java
@@ -1,0 +1,6 @@
+package learn.thyme.thyteams.user;
+
+public enum UserRole {
+    USER,
+    ADMIN
+}