+++ src/main/java/learn/thyme/thyteams/user/web/CreateUserFormData.java
@@ -1,89 +1,42 @@
 package learn.thyme.thyteams.user.web;
 
-import learn.thyme.thyteams.infrastructure.validation.ValidationGroupOne;
 import learn.thyme.thyteams.infrastructure.validation.ValidationGroupTwo;
 import learn.thyme.thyteams.user.CreateUserParameters;
-import learn.thyme.thyteams.user.Gender;
 import learn.thyme.thyteams.user.PhoneNumber;
 import learn.thyme.thyteams.user.UserName;
-import org.springframework.format.annotation.DateTimeFormat;
 
 import javax.validation.constraints.*;
-import java.time.LocalDate;
 
 @NotExistingUser(groups = ValidationGroupTwo.class)
-public class CreateUserFormData {
+@PasswordsMatch(groups = ValidationGroupTwo.class)
+public class CreateUserFormData extends AbstractUserFormData {
     @NotBlank
-    @Size(max = 32, groups = ValidationGroupOne.class)
-    private String firstName;
+    private String password;
     @NotBlank
-    @Size(max = 32, groups = ValidationGroupOne.class)
-    private String lastName;
-    @NotNull
-    private Gender gender;
-    @NotBlank
-    @Email(groups = ValidationGroupOne.class)
-    private String email;
-    @NotNull
-    @DateTimeFormat(pattern = "yyyy-MM-dd")
-    private LocalDate birthday;
-    @NotBlank
-    @Pattern(regexp = "[0-9.\\-() x/+]+", groups = ValidationGroupOne.class)
-    private String phoneNumber;
+    private String passwordRepeated;
 
-    public String getFirstName() {
-        return firstName;
+    public String getPassword() {
+        return password;
     }
 
-    public void setFirstName(String firstName) {
-        this.firstName = firstName;
+    public void setPassword(String password) {
+        this.password = password;
     }
 
-    public String getLastName() {
-        return lastName;
+    public String getPasswordRepeated() {
+        return passwordRepeated;
     }
 
-    public void setLastName(String lastName) {
-        this.lastName = lastName;
-    }
-
-    public Gender getGender() {
-        return gender;
-    }
-
-    public void setGender(Gender gender) {
-        this.gender = gender;
-    }
-
-    public String getEmail() {
-        return email;
-    }
-
-    public void setEmail(String email) {
-        this.email = email;
-    }
-
-    public LocalDate getBirthday() {
-        return birthday;
-    }
-
-    public void setBirthday(LocalDate birthday) {
-        this.birthday = birthday;
-    }
-
-    public String getPhoneNumber() {
-        return phoneNumber;
-    }
-
-    public void setPhoneNumber(String phoneNumber) {
-        this.phoneNumber = phoneNumber;
+    public void setPasswordRepeated(String passwordRepeated) {
+        this.passwordRepeated = passwordRepeated;
     }
 
     public CreateUserParameters toParameters() {
-        return new CreateUserParameters(new UserName(firstName, lastName),
-                gender,
-                birthday,
-                new learn.thyme.thyteams.user.Email(email),
-                new PhoneNumber(phoneNumber));
+        return new CreateUserParameters(new UserName(getFirstName(), getLastName()),
+                password,
+                getGender(),
+                getBirthday(),
+                new learn.thyme.thyteams.user.Email(getEmail()),
+                new PhoneNumber(getPhoneNumber()));
     }
 }